requirements.txt
fastapi
pydantic
uvicorn
pytest
PyYAML
httpx
loguru
watchdog
gemini-webapi
g4f
